feat: #65 카테고리 관리자 인증 추가

+++ src/main/java/com/team2/mosoo_backend/category/controller/CategoryController.java
@@ -5,9 +5,13 @@ import com.team2.mosoo_backend.category.dto.CategoryResponseDto;
 import com.team2.mosoo_backend.category.dto.FirstCategoryResponseDto;
 import com.team2.mosoo_backend.category.dto.SubCategoryResponseDto;
 import com.team2.mosoo_backend.category.service.CategoryService;
+import com.team2.mosoo_backend.exception.CustomException;
+import com.team2.mosoo_backend.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -23,7 +27,13 @@ public class CategoryController {
     // 카테고리 생성
     @PostMapping
     public ResponseEntity<CategoryResponseDto> createCategory(@RequestPart(value = "category") CategoryRequestDto categoryRequestDto,
-                                                 @RequestPart(value = "icon", required = false) MultipartFile file) throws IOException {
+                                                              @RequestPart(value = "icon", required = false) MultipartFile file,
+                                                              @AuthenticationPrincipal UserDetails userDetails) throws IOException {
+
+        if (userDetails.getAuthorities().stream().noneMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"))) {
+            throw new CustomException(ErrorCode.USER_NOT_AUTHORIZED);
+        }
+
         CategoryResponseDto categoryResponseDto = categoryService.createCategory(categoryRequestDto, file);
         return ResponseEntity.status(HttpStatus.OK).body(categoryResponseDto);
     }
@@ -51,14 +61,25 @@ public class CategoryController {
 
     // 카테고리 수정
     @PutMapping("/{categoryId}")
-    public ResponseEntity<CategoryResponseDto> updateCategory(@PathVariable("categoryId") Long categoryId, @RequestBody CategoryRequestDto categoryRequestDto){
+    public ResponseEntity<CategoryResponseDto> updateCategory(@PathVariable("categoryId") Long categoryId, @RequestBody CategoryRequestDto categoryRequestDto,
+                                                              @AuthenticationPrincipal UserDetails userDetails){
+
+        if (userDetails.getAuthorities().stream().noneMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"))) {
+            throw new CustomException(ErrorCode.USER_NOT_AUTHORIZED);
+        }
+
         CategoryResponseDto categoryResponseDto = categoryService.updateCategory(categoryId, categoryRequestDto);
         return ResponseEntity.status(HttpStatus.OK).body(categoryResponseDto);
     }
     
     // 카테고리 삭제
     @DeleteMapping("/{categoryId}")
-    public ResponseEntity<CategoryResponseDto> deleteCategory(@PathVariable("categoryId") Long categoryId){
+    public ResponseEntity<CategoryResponseDto> deleteCategory(@PathVariable("categoryId") Long categoryId, @AuthenticationPrincipal UserDetails userDetails){
+
+        if (userDetails.getAuthorities().stream().noneMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"))) {
+            throw new CustomException(ErrorCode.USER_NOT_AUTHORIZED);
+        }
+
         CategoryResponseDto categoryResponseDto = categoryService.deleteCategory(categoryId);
         return ResponseEntity.status(HttpStatus.OK).body(categoryResponseDto);
     }
